Extracted common AFP functionality to mixin
