+++ lib/msf/core/exploit/mixins.rb
@@ -56,6 +56,7 @@ require 'msf/core/exploit/postgres'
 require 'msf/core/exploit/vim_soap'
 require 'msf/core/exploit/wdbrpc'
 require 'msf/core/exploit/wdbrpc_client'
+require 'msf/core/exploit/afp'
 
 
 # Telephony
